Add ARM CE support.
This also adds support for static builds (define LUA_FFI_BUILD_AS_DLL to get a
dll on windows, otherwise a static lib is built). msvcbuild.bat still builds a
dll.

The build script does not build for CE as this is non-trivial and very
dependent on which CE profile (or even a custom one). Instead to build on CE,
add generate_call_h.bat as a pre-build event and then build *.c with UNDER_CE
defined, whatever defines windows.h requires, and the msvc include directory.

The arm support is also only little endian for the moment.