ffi.new and ffi.cast now support more than argument for structs and arrays
e.g. ffi.new('int[3]', 1, 2, 3)

This is done by trying to set the value using the constructor argument (if
there is only one) and failing that (or if there is more than one), pack it in
a table and then try again. This could probably be smarter and not create the
temporary table, but would require quite a bit more refactoring.

Fixes #25